reflex: Remove bi-directional grpc consuming
Summary:
This bi-directional grpc consuming feature has proven to be a premature optimisation. A sign of this is that is has seen no adoption by the Luno devs.

It is also does not following the principal: "do one thing and do it well". It combines streaming with cursors updates into a single grpc implementation, this is not very usefull since normal streaming and cursor stores already solve this problem.

Reviewers: #backend, neil, andrewn

Reviewed By: #backend, neil, andrewn

Differential Revision: https://phabricator.corp.luno.com/D37264