reflex/sql: Support overriding inserter
Summary:
This change allows complete control over how events are inserted into the DB.

Also, unexport the `Loader` function type. The inserter and loader types are only used externally when setting functional options. It is not actually required to export the type. Also, the surface area of the API of this package is already large, I dot not want to unnecessarily increase it further.

Reviewers: #backend, neil

Reviewed By: #backend, neil

Subscribers: neil

Differential Revision: https://phabricator.corp.luno.com/D41573