reflex: Revert ConsumerName type to just string
Summary:
The reasoning behind the `ConsumerName` string type was to force menu-style enumeration of consumers for better documentation.

```

const (

  // consumerOne does this and that and more.
  consumerOne ConsumerName = "consumer_one"

  // consumerTwo does something else.
  consumerTwo ConsumerName = "consumer_two"
)
```

This did not happen unfortunately #wishfulthinking.

So now reverting to just strings and leaving the concept and implementation of a consumer registry up to the user.

Note: This is a breaking change, sorry.

Reviewers: #backend, shiraaz, bdutoit, vaughan

Reviewed By: #backend, shiraaz, bdutoit, vaughan

Subscribers: vaughan, shiraaz

Differential Revision: https://phabricator.corp.luno.com/D41580