reflex/rpatterns: Implement batch consumer
Summary: Add a `BatchConsumer` to `rpatterns` package that makes it easy to consume batches of events. It maintains an internal buffer which is asynchronously flushed to the consumer when either the "bufferLen" or "bufferPeriod" is reached. It leverages `rpatterns.AckConsumer` to ack the last event after the batch is processed.

Reviewers: corver

Reviewed By: corver

Differential Revision: https://phabricator.corp.luno.com/D44902