reflex: Add best effort consumer option
Summary:
This adds support for a "best effort consumer" which ignores all errors and always continues to the next event.

A common problem is consumers getting "permanently" stuck on some event due to a logic or data issue. The only way to solve this is to fix the underlying issue. This is the correct approach. The problem is that while the consumer is stuck, new events are not being processed; which can have a severe impact on system availability.

To overcome this, a common pattern is to run two identical consumers on the same stream.
 - "Sweeper" consumer: A normal blocking consumer that gets stuck on errors but consumes each and every event (at least once).
 - "Best effort" consumer: A best effort consumer that will never get stuck on errors and always processes new events (no guarantees are provided).

Permanent errors will cause the sweeper to get stuck, but best effort consumer mitigate the impact on the system by processing new events.
Temporary errors (due to network or database issues) might cause the best effort consumer to skip many events, but the sweeper will process them when the temporary issue is resolved.

This pattern doesn't help much if a permanent error is followed by temporary errors. The temporary errors will only be processed once the permanent error is fixed.

It therefore only provides limited mitigation of consumers getting stuck. Stuck consumers should always be fixed asap.

Reviewers: #pod-cryptosec, corver

Reviewed By: #pod-cryptosec, corver

Subscribers: corver

Differential Revision: https://phabricator.corp.luno.com/D52050