rpatterns: Provide consumer name in jettison log context
Summary: Provide consumer name to any jettison logs in the context of a consumer run. Needed in cases where the same function is reused by multiple consumers, so that stack traces in logs are insufficient to distinguish them. Also generally useful to search any logs containing the name of the consumer.

Reviewers: O50 Backend tech leads, #backend, neil

Reviewed By: O50 Backend tech leads, #backend, neil

Differential Revision: https://phabricator.corp.luno.com/D84882