reflex: Provide event id in jettison log context
Summary:
I recently made a mistake when splitting an event consumer into 2 parallel consumers, which we did for performance reasons. The new one (B) was configured to start from the head of the event stream, while the existing one (A) already had a cursor. The result was that some events were left unprocessed (the ones A was meant to process that occurred between the slightly lagging cursor position of A and the latest event when B started).

We had some logs, but it was unnecessarily difficult to work out exactly which events were missed since the event id was not logged. Therefore I propose that reflex append the event id and foreign id to the logging context. I think this is very useful information at a relatively low overhead.

Reviewers: neil

Reviewed By: neil

Subscribers: ahicks

Differential Revision: https://phabricator.corp.luno.com/D90634