rpatterns: Set cursor in primary if not present
Summary: The ReadThroughCursorStore pattern doesn't actually migrate any cursors. You have to wait for the cursor to be set for it to be written to the primary. If you have infrequent events, you may sit around for a long time waiting for the migration to complete. This speeds that up a bit.