rpatterns: Add parallel consumers with manual acking
Summary:

With this change, `rpatterns.ParallelAck` will allow the user to manually acknowledge the events that are being sharded and consumed in parallel. This allows for more flexible consuming, for example, if the user wants to batch events in memory before processing.

I'm sure there's a way to make the various reflex patterns more composable, but I think it would require the "default" consumer implementation to be a manually acking one, as it's the more general pattern. Automatic acking could then be an option or a thin wrapper around the default manually acking consumer. A refactor like above would require some more work, and possibly affect the existing reflex interface, so not within scope of this change.