Add the ability to stream noop events
Summary: Some consumers might require the ability to receive noop events. Receiving these events will be helpful when trying to maintain a clear order of processing...meaning that if I just processed event ID 1....I want to definitely now process event ID 2 regardless if it's a noop event. That way the consumer knows it's always processed all events.

Without this, you have to assume you processed all events and any gap in IDs will just be a result of normal gaps...but it could be the result of something more sinister