internal/metrics: Reduce the number of buckets in histogram metrics
Summary: I've noticed that `reflex_consumer_event_age_seconds_bucket` and `reflex_consumer_latency_seconds_bucket` are the two metrics with the highest cardinality in our production prometheus. I don't think we need so many buckets, so this might be a quick way to reduce the number of time series we're storing.