reflex: Provide a dead letter consumer
Summary: Adds a new dead letter consumer. This is a stage one and the consumer acts like a best efforts consumer but instead of eventually just ignoring the error after the maximum retries it write to an event error table. Further stages will deal with handling and recovery of the errors in the event table - including reply and console lead recovery after data changes (the errors tables records the error message and the id of the event that failed enabling the data to be reconstructed given that it also records the consumer name as well). At this stage it is agnostic as to where the error table lives (it just needs a db reference). Any attempts to provide processing groups to ensure any error for a given (sequential) group also blocking subsequent events until errors are cleared is out of scope as well for this initial work but will probably require changes to the persisted data.

It terms of alternatives to this we do have the pattern of combining a best-efforts consumer with a sweeper which may be considered as an alternative to using this type of a consumer. However that combination cannot really handle:

1. It needs a status change to allow the sweeper to filter out the events that have been handled by the best efforts consumer
2. It requires the sweeper (in general) to lag behind the best efforts consumer, and retrying on the best efforts consumer is thus more likely to collide with the sweeper processing (which is why we tend not to make the best efforts retry at all). All of which is handleable but means that small intermittent errors are more likely to suffer large delays.
3. It means that one poison pill message can block all other errors that may just have been intermittent and there is no process other than changing the code or altering the data after the causal event has been identified from logs to clear it from the sweeper (the error table provides this information so if nothing else the dead letter consumer can act as a better reporting best efforts consumer)
4. They do not provide us with a mechanism in general to expose the problem that are blocking our consumers and although it has not been added yet the error table from dead letter provides a convenient source for a generic replay consumer that can keep "sweeping" across all the clearing all those that can be unblocked as a background task
5. It does not provide the basis for allowing our internal users to self serve clearing issue that the dead letter enables going forward.
6. Somehow even with this in place we still seem to be losing data events!